Allows a second format for URLs
This is to make it easier to use Resque-Brain with Heroku's addon
attachments feature.   Because of how that works, Heroku appends `_URL`
to addon urls that are shared.  So, our current convention of append the
resque name onto them won't work.

But, we don't want to break backwards compatibility, so we allow both.