rework this to make it more reusable at the right level
This was ported from rake tasks which were infinitely flexible, and the
MonitorJob wasn't quite the right abstraction to be able to swap out
the checker/notifier thing.  This makes that possible.

The test is probably more than it needs to be but should mimic the
configuration in the YAML file so we have some confidence there.